% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/as.time.r
\name{as.time}
\alias{as.time}
\title{Convert a vector of frequencies into a vector of times.}
\usage{
as.time(freq)
}
\arguments{
\item{freq}{Vector of frequencies.}
}
\value{
Vector of times.
}
\description{
Convert a vector of frequencies into a vector of times.
}
\examples{
as.time(c(-3, -2, -1, +1, +2, +3))

}

