% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/damped.random.walk.r
\name{damped.random.walk}
\alias{damped.random.walk}
\title{Generate a damped random walk power spectrum.}
\usage{
damped.random.walk(N, freq.max = 1/(2 * pi), tau = 1, SF = 1, sd = 1)
}
\arguments{
\item{N}{Number of frequencies (must be even).}

\item{freq.max}{Value of largest frequency.}

\item{tau}{Characteristic timescale.}

\item{SF}{Value of structure function over long time scales.}

\item{sd}{Standard deviation.}
}
\value{
Frequencies and power spectral density.
}
\description{
\deqn{PSD(f) = \frac{\tau^2 SF_\infty^2}{1 + (2 \pi f \tau)^2}}
}
\examples{
damped.random.walk(100)                # 100 samples with default options
damped.random.walk(100, freq.max = 50) # provide maximum frequency
damped.random.walk(100, sd = 2.0)      # provide standard deviation
damped.random.walk(100, tau = 2.0)     # provide characteristic timescale
damped.random.walk(100, SF = 2.0)      # provide value of structure function

}

