% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/freq.r
\name{freq}
\alias{freq}
\title{Generate a vector of frequencies.}
\usage{
freq(N, freq.max = 1/(2 * pi))
}
\arguments{
\item{N}{Number of elements in vector (must be even).}

\item{freq.max}{Value of largest frequency.}
}
\value{
Vector of frequencies.
}
\description{
Generate a vector of frequencies.
}
\examples{
freq(10, freq.max = 5) # => c(-5, -4, -3, -2, -1, 1, 2, 3, 4, 5)

}

