% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mirror.complex.r
\name{mirror.complex}
\alias{mirror.complex}
\title{Takes a complex valued vector and returns a new vector, which is
mirrored at index 0, but with the mirrored values having opposite
imaginary parts.}
\usage{
mirror.complex(z)
}
\arguments{
\item{z}{Complex valued vector.}
}
\value{
Mirrored vector.
}
\description{
TODO: Explain this better.
}
\examples{
mirror.complex(c(1+1i, 2+2i)) # => c(2-2i, 1-1i, 1+1i, 2+2i)

}

