% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/noisy.power.spectrum.r
\name{noisy.power.spectrum}
\alias{noisy.power.spectrum}
\title{Takes a power spectrum and makes it noisy, by adding white noise to
its amplitude, and introducing random, uniformly distributed
phases.}
\usage{
noisy.power.spectrum(power, sd = 1)
}
\arguments{
\item{power}{Power spectral density.}

\item{sd}{Standard deviation of white noise.}
}
\value{
Power with noise added.
}
\description{
\deqn{A(f) \sim \mathcal{N}(PSD(f), \sigma^2)}
\deqn{\phi(f) \sim \mathcal{U}(0, 2\pi)}
\deqn{z(f) = A(f) \exp(i \phi(f))}
}

