% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.spectrum.r
\name{plot.spectrum}
\alias{plot.spectrum}
\title{Plots a power spectrum and time series side-by-side.}
\usage{
\method{plot}{spectrum}(power.spectrum, time.series,
  output = "spectrum_plot.pdf")
}
\arguments{
\item{power.spectrum}{Data frame containing freq and power.}

\item{time.series}{Data frame containing time and signal.}
}
\description{
Plots a power spectrum and time series side-by-side.
}

