% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/time.series.r
\name{time.series}
\alias{time.series}
\title{Transform a power spectrum into a time series, using an inverse
fast Fourier transform.}
\usage{
\method{time}{series}(freq, power)
}
\arguments{
\item{freq}{The frequency.}

\item{power}{The power spectral distribution.}
}
\value{
Time series.
}
\description{
Transform a power spectrum into a time series, using an inverse
fast Fourier transform.
}

